pyproject.toml
[tool.poetry]
name = "tuxseo"
version = "0.1.0"
description = "Best SEO Blog Writer entity in the universe!"
authors = ["Rasul Kireev <kireevr1996@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
dj-stripe = "^2.9.0"
django-allauth = {extras = ["socialaccount"], version = "^64.0.0"}
django-anymail = {extras = ["mailgun"], version = "^12.0"}
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-ninja = "^1.3.0"
django-q2 = {extras = ["sentry"], version = "^1.8.0"}
django-storages = {extras = ["s3"], version = "^1.14.4"}
django-structlog = "^8.1.0"
django-widget-tweaks = "^1.4.12"
gunicorn = "^23.0.0"
ipython = "^8.27.0"
logfire = "^3.6.4"
markdown = "^3.7"
psycopg2 = "^2.9.9"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
python-webpack-boilerplate = "^1.0.0"
redis = "^5.0.8"
stripe = "^11.6.0"
structlog = "^24.4.0"
whitenoise = "^6.7.0"
posthog = "^5.3.0"
sentry-sdk = {extras = ["pydantic-ai"], version = "^2.45.0"}
structlog-sentry = "^2.2.1"
django-mjml = "^1.4"
weasyprint = "^66.0"
pydantic-ai = "^1.9.1"
python-frontmatter = "^1.1.0"
pyyaml = "^6.0.3"
pygments = "^2.19.2"
replicate = "^1.0.7"
pgvector = "^0.4.1"
gpt-researcher = "^0.14.5"
pillow = "^12.0.0"
exa-py = "^2.0.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.1"
djlint = "^1.36.4"
ruff = "^0.12.0"
ty = "^0.0.1a11"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "B",    # flake8-bugbear
    "I",    # isort
    "DJ",   # django-specific rules
    "UP",   # pyupgrade
    "C90",  # mccabe complexity
]
ignore = ["E402"]

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["E501", "F401", "F403", "F405"]
"**/models.py" = ["C901"]

[tool.ruff.format]
quote-style = "double"

[tool.djlint]
profile="django"
ignore = "H031,H006,H023,H021,H011,T002"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
